feat: Define and annotate the following variables with the specified values:

+++ python_variable_annotations/4-define_variables.py
@@ -9,7 +9,7 @@ def floor(n: float) -> int:
     return math.floor(n)
 
 
-a: float = 1
+a: int = 1
 pi: float = 3.14
 i_understand_annotations: bool = True
 school: str = "Holberton"
